Appropriately setup an inpm project with eleventy
I was lazy before this and didn't setup the project. Now, the project is
a proper npm project so that it is easier to start making changes.

Now we can just npm install and npm start to get going and the required
configuration is there to not overwrite stuff.

There is one meaningful break in this change. Previously, the project
was building to root (and served from root). Now it builds to /docs. So,
GH Pages needs to be reconfigured to serve from /docs.